fix: fix handling of lambda functions
