+++ src/components/chatgpt-waf-log-processor.ts
@@ -1,5 +1,3 @@
-
-import { PythonFunction } from '@aws-cdk/aws-lambda-python-alpha';
 import * as cdk from 'aws-cdk-lib';
 import * as dynamodb from 'aws-cdk-lib/aws-dynamodb';
 import { Rule, Schedule } from 'aws-cdk-lib/aws-events';
@@ -8,7 +6,7 @@ import * as iam from 'aws-cdk-lib/aws-iam';
 import * as lambda from 'aws-cdk-lib/aws-lambda';
 //import * as logs from 'aws-cdk-lib/aws-logs'
 //import * as cw from 'aws-cdk-lib/aws-cloudwatch'
-import { Architecture, Runtime } from 'aws-cdk-lib/aws-lambda';
+import { Architecture } from 'aws-cdk-lib/aws-lambda';
 import { Construct } from 'constructs';
 
 export interface IChatGPTWafLogProcessorProps {
@@ -58,30 +56,23 @@ export class ChatGPTWafLogProcessor extends Construct {
         actions: ['wafv2:*'],
       }),
     );
-    const waf_log_processor_lambda = new PythonFunction(
-      this,
-      'waf-log-check-lambda',
-      {
-        entry: './src/components/lambda/chatgpt_result_processor/',
-        runtime: new lambda.Runtime(
-          Runtime.PYTHON_3_10.name,
-          lambda.RuntimeFamily.PYTHON,
-        ),
-        role: waf_log_processor_lambda_role,
-        description:
-            'Reads DynamoDB where IPs are listed by ChatBGP to block and blocks them bei syncing to an IPSet.',
-        index: 'result_processor.py',
-        architecture: Architecture.ARM_64,
-        handler: 'handler',
-        environment: {
-          PRODUCT: 'ChatGPTBadIPs',
-          DB_NAME: table.tableName,
-          IP_SET_NAME: this.ip_set_name,
-          SCOPE: this.rule_scope,
-        },
-        timeout: cdk.Duration.seconds(20),
+    const waf_log_processor_lambda = new lambda.Function(this, 'waf-log-process-lambda', {
+      runtime: lambda.Runtime.PYTHON_3_10,
+      code: lambda.Code.fromAsset('./src/components/lambda/chatgpt_result_processor.zip', {
+      }),
+      handler: 'result_processor.handler',
+      role: waf_log_processor_lambda_role,
+      description:
+          'Reads DynamoDB where IPs are listed by ChatBGP to block and blocks them bei syncing to an IPSet.',
+      architecture: Architecture.ARM_64,
+      environment: {
+        PRODUCT: 'ChatGPTBadIPs',
+        DB_NAME: table.tableName,
+        IP_SET_NAME: this.ip_set_name,
+        SCOPE: this.rule_scope,
       },
-    );
+      timeout: cdk.Duration.seconds(20),
+    });
     table.grantWriteData(waf_log_processor_lambda);
     table.grantReadData(waf_log_processor_lambda);
     new Rule(this, 'waf-chatgpt-analysis-rule', {
