+++ src/components/lambda/log_analytics/requirements.txt
@@ -1,0 +1,2 @@
+boto3 
+openai
