--- src/components/lambda/waf_logs_chatgpt/.env
@@ -1,4 +1,0 @@
-export PRODUCT="ChatGPTBadIPs"
-export DB_NAME="waf_ip_protection"
-export SECRET_ID="ChatGPT_API_Secret"
-export SCOPE="REGIONAL"
