--- src/components/lambda/waf_logs_chatgpt/requirements.txt
@@ -1,2 +1,0 @@
-boto3 
-openai
